For goproxy can be compatible for go 1.11 and go 1.12, modify generate script.
